Move tours kicker out of onload.js; just use the import.
